+++ static/js/src/connected-elements/Level.vue
