+++ static/js/src/connected-elements/TextReadOnly.vue
@@ -10,7 +10,7 @@
 </template>
 
 <script>
-import mapValueToElement from '../mixins/mapValueToElement';
+import mapValueToElement from './mixins/mapValueToElement';
 
 export default {
   name: 'TextInput',
