+++ static/js/src/connected-elements/mixins/updateSingleValueOnInput.js
@@ -1,0 +1,13 @@
+export default {
+  methods: {
+    onInput(newValue) {
+      this.$store.dispatch(
+        `${this.storeModule}/updateValue`,
+        { key: this.name, value: newValue },
+      );
+
+      this.$emit('updateCallback', newValue);
+      if (this.validator) this.validate(newValue);
+    },
+  },
+};
