+++ static/js/src/mixins/form/validateOnInput.js
@@ -16,16 +16,6 @@ export default {
   },
 
   methods: {
-    onInput(newValue) {
-      this.$store.dispatch(
-        `${this.storeModule}/updateValue`,
-        { key: this.name, value: newValue },
-      );
-
-      this.$emit('updateCallback', newValue);
-      if (this.validator) this.validate(newValue);
-    },
-
     validate(value) {
       let message = '';
